Change coc-config for new pum
CoC changed to a pum which is shown in a popup menu. Some changes are
needed to preserve the old vim behaviour, since the new pum is more
oriented towards VSCode behaviour.

Adding `suggest.selection` with `none`, so that nothing is selected by
default. Also add some more useful options.

Signed-off-by: SteveCharleston <steven@schalhorn.org>