Add second compl trigger for nested completitions
When using `<C-j>` as completition trigger, it is not possible to start
another completitions with jumpable targets, since we would always jump
out of the first completition item. Only when using another trigger we
can also expand snippets inside an ongoing jumpable completition.

It is unknown why this doesn't work otherwise but it is a sufficient
workaround.

Signed-off-by: Steven Schalhorn <steven.schalhorn@ionos.com>