vim: Fix gruvbox colors after latest update
Gruvbox changed *all* colors around in an effort to make coloring
between languages more consistent. This had the unfortunate effect that
the coloring of languages was completly different to what it was before.

Fortunately an option `gruvbox_legacy_language_groups` had been
introduced to restore the old coloring.

As an additional change the coloring of has been made explicitely,
because the coloring of CoCs floating windows does not work as intendent
in recent versions of neovim.

Signed-off-by: SteveCharleston <steven@schalhorn.org>